extract files and components
