+++ frontend/.gitignore
@@ -21,5 +21,3 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-
-src/mock-data.json
