+++ frontend/src/data/types.ts
@@ -1,0 +1,18 @@
+export type Activity = {
+  name: string;
+  active: boolean;
+};
+
+export type ActivityRecord = {
+  date: string;
+  activity: Activity;
+};
+
+export type ActivitySummary = {
+  count: number;
+  active: boolean;
+};
+
+export type ActivitySummaryMap = {
+  [key: string]: ActivitySummary;
+};
